cambio de tipo de variable

+++ src/main/java/com/nttdata/accounts/entity/BankAccount.java
@@ -19,7 +19,7 @@ public class BankAccount {
     private String numberAccount; //NúmeroDeCuenta
     private String idCustomer; //Id del cliente
     private String codProfile; //Codigo de perfil del cliente
-    private Float balance; //Saldo
+    private Double balance; //Saldo
     @DateTimeFormat(pattern = "yyyy-mm-dd")
     private Date date;
     private Boolean status;
